Split transaction.rs into submodules.
